出差情况功能设计1

+++ src/main/java/com/iwork/service/ClockinServiceImpl.java
@@ -9,8 +9,12 @@ import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
+import java.util.logging.SimpleFormatter;
+
 @Service
 public class ClockinServiceImpl implements ClockinService{
     @Autowired
@@ -44,7 +48,7 @@ public class ClockinServiceImpl implements ClockinService{
             vocation.setUsername(user.selectByUserId(vocation.getUserId()).getUserName());
             vocation.setSubperiorname(user.selectByUserId(vocation.getSubperior()).getUserName());
             if (!vocation.getStatus().equals(type)){
-                if (type==4&&vocation.getStatus().equals(5))
+                if (type==4 && vocation.getStatus().equals(5))
                     break;
                 it.remove();
 
@@ -58,6 +62,7 @@ public class ClockinServiceImpl implements ClockinService{
     @Override
     public Boolean insertIntoVocation(Vocation record) {
         logger.info("1");
+        record.setStatus(4);
         vocationMapper.insert(record);
 
         return true;
@@ -65,8 +70,21 @@ public class ClockinServiceImpl implements ClockinService{
 
     @Override
     public Boolean deleteVocationById(int vid) {
+        Vocation vocation =vocationMapper.selectByPrimaryKey(vid);
 
-        vocationMapper.deleteByPrimaryKey(vid);
+
+        if (vocation.getStatus()==2)
+        {
+            Date now =new Date();
+            SimpleDateFormat sdf =new SimpleDateFormat("yyyy-MM-dd HH:mm");
+
+            vocation.setStatus(3);
+            vocation.setEndTime(sdf.format(now));
+            vocationMapper.updateByPrimaryKey(vocation);
+        }
+        else
+            vocationMapper.deleteByPrimaryKey(vid);
+
         return true;
     }
 
@@ -99,7 +117,38 @@ public class ClockinServiceImpl implements ClockinService{
         return vocationMapper.updateByPrimaryKeySelective(record);
     }
 
-    private void setVocationStatus(Vocation v){
+    @Override
+    public void UpdateStatus() {
+        List<Vocation> vocationList;
+        Date day =new Date();
+        SimpleDateFormat sdf= new SimpleDateFormat("yyyy-MM-dd HH:mm");
+        Date starttime;
+        Date endtime;
+        String timeNow =sdf.format(day);
+        for (int i=1;i<3;i++)
+        {
+        vocationList = vocationMapper.updateRecordStatus(i);
+        Iterator<Vocation> it =vocationList.iterator();
+        while (it.hasNext()){
+            Vocation vocation =it.next();
+            try{
+                int s=vocation.getStatus();
+                starttime =sdf.parse(vocation.getStartTime());
+                endtime =sdf.parse(vocation.getEndTime());
+                Date now =sdf.parse(timeNow);
+                if (now.before(starttime))
+                    vocation.setStatus(1);
+                else if (now.before(endtime))
+                    vocation.setStatus(2);
+                else if (now.after(endtime))
+                    vocation.setStatus(3);
+                if (s!=vocation.getStatus())
+                    vocationMapper.updateByPrimaryKeySelective(vocation);
+        }catch (Exception e){
+                e.printStackTrace();
+            }
 
+        }
+        }
     }
 }
